Remove legacy ops
