+++ dali/operators/decoder/video/video_decoder_mixed.cc
@@ -1,4 +1,4 @@
-// Copyright (c) 2022-2024, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// Copyright (c) 2022, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,23 +12,46 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <vector>
-#include "dali/pipeline/operator/operator.h"
+#include "dali/operators/decoder/video/video_decoder_mixed.h"
 
 namespace dali {
 
-class VideoDecoderMixed : public Operator<MixedBackend> {
- public:
-  explicit VideoDecoderMixed(const OpSpec &spec): Operator<MixedBackend>(spec) {
-    DALI_FAIL(
-        "Video operators are now part of a separate package. "
-        "Please use the operators from module `nvidia.dali.fn.plugin.video` instead");
+bool VideoDecoderMixed::SetupImpl(
+  std::vector<OutputDesc> &output_desc, const Workspace &ws) {
+  ValidateInput(ws);
+  const auto &input = ws.Input<CPUBackend>(0);
+  int batch_size = input.num_samples();
+  auto stream = ws.stream();
+  frames_decoders_.resize(batch_size);
+  for (int i = 0; i < batch_size; ++i) {
+    thread_pool_.AddWork([this, i, &input, stream](int) {
+      auto sample = input[i];
+      auto data = reinterpret_cast<const char *>(sample.data<uint8_t>());
+      size_t size = sample.shape().num_elements();
+      frames_decoders_[i] = std::make_unique<FramesDecoderGpu>(data, size, stream, false);
+    });
   }
+  thread_pool_.RunAll();
+  output_desc.resize(1);
+  output_desc[0].shape = ReadOutputShape();
+  output_desc[0].type = DALI_UINT8;
+  return true;
+}
 
-  bool CanInferOutputs() const override { return true; }
-  void Run(Workspace &ws) override { assert(false); }
-  bool SetupImpl(std::vector<OutputDesc> &output_desc, const Workspace &ws) override { assert(false); }
-};
+void VideoDecoderMixed::Run(Workspace &ws) {
+  auto &output = ws.Output<GPUBackend>(0);
+  const auto &input = ws.Input<CPUBackend>(0);
+  int batch_size = input.num_samples();
+  for (int s = 0; s < batch_size; ++s) {
+    thread_pool_.AddWork([this, s, &output](int) {
+      DecodeSample(output[s], s);
+      // when the decoding is done release the decoder,
+      // so it can be reused by the next sample in the batch
+      frames_decoders_[s].reset();
+    }, input[s].shape().num_elements());
+  }
+  thread_pool_.RunAll();
+}
 
 DALI_REGISTER_OPERATOR(experimental__decoders__Video, VideoDecoderMixed, Mixed);
