+++ plugins/dali_video/src/nvidia/dali_video/CMakeLists.txt
@@ -13,7 +13,6 @@
 # limitations under the License.
 
 cmake_minimum_required(VERSION 3.21)
-#set(CMAKE_CUDA_ARCHITECTURES "50;60;70;80;90")
 project(dali_video LANGUAGES CUDA CXX C)
 
 set(CMAKE_CXX_STANDARD 17)
@@ -68,21 +67,12 @@ endif()
 message(STATUS "DALI_LIB_DIR=${DALI_LIB_DIR}")
 
 set(VIDEO_PLUGIN_SOURCES
-    dynlink_nvcuvid/dynlink_nvcuvid_gen.cc  # TODO(janton): remove
-    dynlink_nvcuvid/dynlink_nvcuvid.cc      # TODO(janton): remove
-    loader/frames_decoder.cc
-    loader/frames_decoder_gpu.cc
-    loader/video_loader_decoder_cpu.cc
-    loader/video_loader_decoder_gpu.cc
-    input/video_input_cpu.cc
-    input/video_input_mixed.cc
-    decoder/video_decoder_cpu.cc
     decoder/video_decoder_mixed.cc
-    decoder/new_video_decoder_mixed.cc
 )
 
 add_library(dali_video SHARED ${VIDEO_PLUGIN_SOURCES})
 set_target_properties(dali_video PROPERTIES POSITION_INDEPENDENT_CODE ON)
+set_target_properties(dali_video PROPERTIES CUDA_ARCHITECTURES OFF)
 target_include_directories(dali_video PUBLIC ${dali_video_SRC_DIR})
 target_link_directories(dali_video PUBLIC ${DALI_LIB_DIR})
 target_link_libraries(dali_video PUBLIC dali dali_core dali_kernels VideoCodecSDKUtils)
