--- plugins/dali_video/src/nvidia/dali_video/dynlink_nvcuvid/dynlink_nvcuvid.h
@@ -1,25 +1,0 @@
-// Copyright (c) 2017-2022, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#pragma once
-
-#include <string>
-#include "dali/core/cuda_error.h"
-#include "nvcuvid.h"
-
-#define NVCUVID_CALL(arg) CUDA_CALL(arg)
-#define NVCUVID_API_EXISTS(arg) (cuvidIsSymbolAvailable(#arg))
-
-bool cuvidInitChecked(void);
-bool cuvidIsSymbolAvailable(const char *name);
