--- plugins/dali_video/src/nvidia/dali_video/loader/nvdecode/color_space.h
@@ -1,27 +1,0 @@
-// Copyright (c) 2022, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#pragma once
-
-#include <stdint.h>
-
-void yuv_to_rgb(
-    uint8_t *yuv,
-    int yuv_pitch,
-    uint8_t *rgb,
-    int rgb_pitch,
-    int width,
-    int height,
-    bool full_range,
-    cudaStream_t stream);
